build2

+++ resources/js/Components/Helpers.js
@@ -12,11 +12,11 @@ export function formatPrice(amount, currency = 'USD') {
         currency,
     }).format(amount);
 }
-export function formatDateTime(date, options = { year: 'numeric', month: 'long', day: 'numeric', hour: '2-digit', minute: '2-digit' }) {
+export function formatDateTime(date, options = { year: 'numeric', month: 'short', day: 'numeric', hour: '2-digit', minute: '2-digit' }) {
     return new Intl.DateTimeFormat('en-US', options).format(new Date(date));
 }
 
-export function formatDate(date, options = { year: 'numeric', month: 'long', day: 'numeric' }) {
+export function formatDate(date, options = { year: 'numeric', month: 'short', day: 'numeric' }) {
     return new Intl.DateTimeFormat('en-US', options).format(new Date(date));
 }
